@@ -16,8 +16,8 @@
     "@types/markdown-it": "^12.2.3",
     "@types/markdown-it-container": "^2.0.5",
     "@types/prettier": "^2.7.1",
-    "@types/react": "^17.0.24",
-    "@types/react-dom": "^17.0.9",
+    "@types/react": "^18.0.21",
+    "@types/react-dom": "^18.0.6",
     "@vuepress/client": "^2.0.0-beta.51",
     "@vuepress/plugin-register-components": "^2.0.0-beta.51",
     "@vuepress/plugin-search": "^2.0.0-beta.51",
@@ -25,8 +25,8 @@
     "front-matter": "^4.0.2",
     "markdown-it-container": "^3.0.0",
     "prettier": "^2.7.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "typedoc": "^0.23.15",
     "typedoc-plugin-markdown": "^3.13.6",
     "typescript": "^4.7.2",
