@@ -22,8 +22,6 @@
   },
   "devDependencies": {
     "@babel/core": "^7.18.13",
-    "@docusaurus/module-type-aliases": "2.1.0",
-    "@docusaurus/theme-classic": "2.1.0",
     "@dxos/async": "workspace:*",
     "@dxos/esbuild": "workspace:*",
     "@dxos/esbuild-plugins": "~2.29.0",
@@ -63,7 +61,7 @@
     "@types/lodash.set": "^4.3.7",
     "@types/mocha": "^8.2.2",
     "@types/node": "^16.11.58",
-    "@types/react": "^17.0.24",
+    "@types/react": "^18.0.21",
     "@types/sharedworker": "0.0.80",
     "@types/uuid": "~8.3.3",
     "@typescript-eslint/eslint-plugin": "^5.29.0",
@@ -91,8 +89,8 @@
     "pkg-up": "^3.1.0",
     "playwright": "^1.25.1",
     "raf": "^3.4.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "require-from-string": "^2.0.2",
     "rollup": "~2.77.3",
     "syncpack": "^8.2.4",
