@@ -25,8 +25,8 @@
     "@mui/icons-material": "^5.8.0",
     "@mui/material": "^5.8.1",
     "debug": "^4.3.3",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-router-dom": "^6.3.0"
   },
   "devDependencies": {
@@ -35,8 +35,8 @@
     "@dxos/vite-plugin": "workspace:*",
     "@types/debug": "^4.1.7",
     "@types/node": "^16.11.58",
-    "@types/react": "^17.0.24",
-    "@types/react-dom": "^17.0.9",
+    "@types/react": "^18.0.21",
+    "@types/react-dom": "^18.0.6",
     "@vitejs/plugin-react": "^2.0.1",
     "typescript": "^4.7.2",
     "vite": "3.0.9",
