@@ -36,9 +36,9 @@
     "color-hash": "^1.0.3",
     "debug": "^4.3.3",
     "moment": "^2.24.0",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-copy-to-clipboard": "^5.1.0",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "webextension-polyfill": "^0.8.0"
   },
   "devDependencies": {
@@ -47,8 +47,8 @@
     "@types/chrome": "^0.0.125",
     "@types/debug": "^4.1.7",
     "@types/download": "^8.0.1",
-    "@types/react": "^17.0.24",
-    "@types/react-dom": "^17.0.9",
+    "@types/react": "^18.0.21",
+    "@types/react-dom": "^18.0.6",
     "@types/webextension-polyfill": "^0.8.0",
     "body-parser": "1.19.0",
     "chalk": "^4.1.0",
