@@ -33,10 +33,10 @@
     "@dxos/esbuild-plugins": "~2.29.0",
     "@dxos/protocols": "workspace:*",
     "@dxos/protocols-toolchain": "workspace:*",
-    "@types/react": "^17.0.24",
+    "@types/react": "^18.0.21",
     "@types/react-copy-to-clipboard": "~5.0.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-router-dom": "^6.3.0"
   },
   "peerDependencies": {
