@@ -43,9 +43,9 @@
     "color-hash": "^1.0.3",
     "moment": "^2.24.0",
     "protobufjs": "^6.9.0",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-copy-to-clipboard": "^5.1.0",
-    "react-dom": "^17.0.2"
+    "react-dom": "^18.2.0"
   },
   "devDependencies": {
     "@babel/core": "^7.18.13",
@@ -56,9 +56,9 @@
     "@mui/system": "^5.8.1",
     "@types/color-hash": "^1.0.1",
     "@types/faker": "^5.5.9",
-    "@types/react": "^17.0.24",
+    "@types/react": "^18.0.21",
     "@types/react-copy-to-clipboard": "~5.0.1",
-    "@types/react-dom": "^17.0.9",
+    "@types/react-dom": "^18.0.6",
     "faker": "^5.5.3"
   }
 }
