@@ -53,7 +53,7 @@
     "lodash.defaultsdeep": "^4.6.1",
     "node-fetch": "^2.6.0",
     "oclif": "^3.1.2",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "ws": "^7.4.4"
   },
   "devDependencies": {
