@@ -3,7 +3,7 @@
 //
 
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 
 import { Event } from '@dxos/async';
 import { Client } from '@dxos/client';
@@ -11,8 +11,6 @@ import { Client } from '@dxos/client';
 import { App } from './App';
 
 export const initializeDevtools = (clientReady: Event<Client>) => {
-  ReactDOM.render(
-    <App clientReady={clientReady} />,
-    document.getElementById('root')
-  );
+  createRoot(document.getElementById('root')!)
+    .render(<App clientReady={clientReady} />);
 };
