@@ -3,7 +3,7 @@
 //
 
 import React from 'react';
-import { render } from 'react-dom';
+import { createRoot } from 'react-dom/client';
 // import debug from 'debug'
 // import { log } from '@dxos/log'
 
@@ -13,9 +13,9 @@ import { render } from 'react-dom';
 import { App } from './App';
 
 // TODO(wittjosiah): StrictMode causing issues with the react sdk, re-enable once fixed.
-render(
-  // <StrictMode>
-  <App />,
-  // </StrictMode>
-  document.getElementById('root')
-);
+createRoot(document.getElementById('root')!)
+  .render(
+    // <StrictMode>
+    <App />
+    // </StrictMode>
+  );
