@@ -3,7 +3,7 @@
 //
 
 import React from 'react';
-import { render } from 'react-dom';
+import { createRoot } from 'react-dom/client';
 
 import { Config, Defaults, Dynamics } from '@dxos/config';
 import { ClientProvider } from '@dxos/react-client';
@@ -14,13 +14,13 @@ import { App } from './App';
 const configProvider = async () => new Config(await Dynamics(), Defaults());
 
 (() => {
-  render(
-    <ClientProvider config={configProvider}>
-      {/* TODO(wittjosiah): Remove once HALO app integration works. */}
-      <ProfileInitializer>
-        <App />
-      </ProfileInitializer>
-    </ClientProvider>,
-    document.getElementById('root')
-  );
+  createRoot(document.getElementById('root')!)
+    .render(
+      <ClientProvider config={configProvider}>
+        {/* TODO(wittjosiah): Remove once HALO app integration works. */}
+        <ProfileInitializer>
+          <App />
+        </ProfileInitializer>
+      </ClientProvider>
+    );
 })();
