@@ -3,7 +3,7 @@
 //
 
 import React, { StrictMode, useState } from 'react';
-import { render } from 'react-dom';
+import { createRoot } from 'react-dom/client';
 
 import { schema } from '@dxos/protocols';
 import { useAsyncEffect } from '@dxos/react-async';
@@ -85,18 +85,18 @@ if (typeof SharedWorker !== 'undefined') {
     });
     await channel.addPort(worker.port);
 
-    render(
-      <StrictMode>
-        <div style={{
-          display: 'flex'
-        }}>
-          {ports!.map(port => (
-            <App key={port.id} {...port} />
-          ))}
-        </div>
-      </StrictMode>,
-      document.getElementById('root')
-    );
+    createRoot(document.getElementById('root')!)
+      .render(
+        <StrictMode>
+          <div style={{
+            display: 'flex'
+          }}>
+            {ports!.map(port => (
+              <App key={port.id} {...port} />
+            ))}
+          </div>
+        </StrictMode>
+      );
   })();
 } else {
   throw new Error('Requires a browser with support for shared workers.');
