@@ -3,7 +3,7 @@
 //
 
 import React from 'react';
-import { render } from 'react-dom';
+import { createRoot } from 'react-dom/client';
 
 import { App } from './components';
 
@@ -21,8 +21,6 @@ if ('serviceWorker' in navigator) {
 }
 
 (() => {
-  render(
-    <App />,
-    document.getElementById('root')
-  );
+  createRoot(document.getElementById('root')!)
+    .render(<App />);
 })();
